Emulator test for NregionCommit Write Barrier
- Used Http interceptor to "mock" nregioncommit feature enabled in AccountProperties.
- Used Transport interceptor to mock GLobalNRegionCommitLSN for testing scenario.
- Tested success and failures cases.

+++ Microsoft.Azure.Cosmos/src/direct/StoreResult.cs
@@ -210,7 +210,7 @@ namespace Microsoft.Azure.Documents
                     headerValue = documentClientException.Headers[WFConstants.BackendHeaders.GlobalNRegionCommittedGLSN];
                     if (!string.IsNullOrEmpty(headerValue))
                     {
-                        globalNRegionCommittedGLSN = int.Parse(headerValue, CultureInfo.InvariantCulture);
+                        globalNRegionCommittedGLSN = long.Parse(headerValue, CultureInfo.InvariantCulture);
                     }
                     ISessionToken sessionToken = null;
                     if (StoreResult.UseSessionTokenHeader)
